Control.Monad.Trans.Fault

Description

monad-fault provides an extensible way to trigger arbitrary failures within your programs. Potential faults are tracked in the type system via MonadFault constraints. Useful for testing fault tolerance (e.g. retry logic)

For instance:

mightFault :: (MonadIO m, MonadFaults '["redis", "s3", "sqs"] m)
           => Redis.Connection
           -> AWS.Env
           -> m ()
mightFault redisConn awsEnv = do
  s3Result <- faulty "s3" $ readFromS3 awsEnv
  faulty "redis" $ writeToRedis redisConn s3Result
  faulty @"sqs" $ writeToSQS awsEnv s3Result

In production, we'd run this with FaultlessT which will cause the faults to be elided:

runFaultlessT (mightFault redisConn awsEnv) :: IO ()

But while testing, we can create a FaultController and run with FaultyT. In our tests, we can set different parts of the program to fault using setFault & resetFault. We can then confirm, for instance, that a Redis blip followed by a retry still results in the correct effects being performed on the world:

fc <- newFaultController '["redis", "s3", "sqs"]
setFault "redis" fc (redisException "uh oh!")
runFaultyT fc (mightFault redisConn awsEnv) shouldThrow anyException
resetFault @"redis" fc
runFaultyT fc (mightFault redisConn awsEnv)
performChecks

Causing faults

fault :: forall fault m. MonadFault fault m => m ()

Cause a fault named fault

Meant to be used w/TypeApplications

>>> fault @"redis" :: MonadFault "redis" m => m ()

faulty :: forall fault m a. MonadFault fault m => m a -> m a

Tag an action as a potential fault named fault

>>> faulty @"failure name" $ mightFail

The MonadFault Class

class Monad m => MonadFault fault m where

m is capable of having fault. faults are named with type-level strings

Minimal complete definition

faultPrx

Methods

faultPrx :: Proxy fault -> m ()

Cause a fault named fault

type family MonadFaults (faults :: [Symbol]) (m :: * -> *) :: Constraint where ...

One m frequently has many potential faults

Equations

MonadFaults '[] m = ()

MonadFaults (fault ': rest) m = (MonadFault fault m, MonadFaults rest m)

The FaultlessT & FaultyT monad transformers

newtype FaultlessT m a

Can never fault.

Constructors

FaultlessT

Fields

unFaultlessT :: IdentityT m a

runFaultlessT :: FaultlessT m a -> m a

Unwrap FaultlessT, ignoring all possible faults

newtype FaultyT faults m a

Monad transformer that allows the caller to control which faults occur

Constructors

FaultyT

Fields

unFaultyT :: ReaderT (FaultController faults) m a

runFaultyT :: FaultController faults -> FaultyT faults m a -> m a

Run a FaultyT, causing faults along the way according to the given FaultController

Controlling faults

newFaultController :: NewFault faults => IO (FaultController faults)

Create a FaultController initially configured to never fault

setFault :: forall fault faults e. (HasFault fault faults, Exception e) => e -> FaultController faults -> IO ()

Set a fault to throw a given Exception

resetFault :: forall fault faults e. (HasFault fault faults, Exception e) => FaultController faults -> IO ()

Set a fault to not fail.

Debugging

printFaultController :: FaultController faults -> IO ()

>>> fc <- newFaultController @'["x", "y"]
>>> printFaultController fc
(FCCons @"x" FaultConfig Nothing (FCCons @"y" FaultConfig Nothing FCNil))

showFaultController :: FaultController faults -> IO String

Create the output of printFaultController

askFaultController :: Monad m => FaultyT faults m (FaultController faults)

Access the FaultController within a FaultyT. You usually can't use this because you'll write your programs in terms of MonadFault "whatever" m =>

Internals

data FaultController faults where

Extensible record of FaultConfigs, each tagged with a fault name at the type level

Constructors

FCNil :: FaultController '[]

FCCons :: forall f rest. KnownSymbol f => Proxy f -> !(IORef FaultConfig) -> FaultController rest -> FaultController (f ': rest)

data FaultConfig

If the exception is Just, we fault. Otherwise, we don't

Constructors

FaultConfig (Maybe SomeException)

class NewFault faults where

Create a default, non-faulting FaultController

Minimal complete definition

newFaultController

Methods

newFaultController :: IO (FaultController faults)

Create a FaultController initially configured to never fault

class HasFault f faults where

Query & modify a FaultController at certain faults

Minimal complete definition

getFaultConfig, setFaultConfig

Methods

getFaultConfig :: FaultController faults -> IO FaultConfig

setFaultConfig :: FaultConfig -> FaultController faults -> IO ()